mcq-generator/requirements.txt
streamlit
langchain
langchain-openai
langchain-google-genai
langchain-community
openai
python-dotenv
pypdf
pandas
jupyter
tiktoken